// src/management/api/errors/ForbiddenError.ts
// This file was auto-generated by Fern from our API Definition.

import * as errors from "../../errors/index.js";
import * as core from "../../core/index.js";

export class ForbiddenError extends errors.ManagementError {
    constructor(body?: unknown, rawResponse?: core.RawResponse) {
        super({
            message: "ForbiddenError",
            statusCode: 403,
            body: body,
            rawResponse: rawResponse,
        });
        Object.setPrototypeOf(this, ForbiddenError.prototype);
    }
}