// src/management/api/errors/NotFoundError.ts
// This file was auto-generated by Fern from our API Definition.

import * as core from "../../core/index.js";
import * as errors from "../../errors/index.js";

export class NotFoundError extends errors.ManagementError {
    constructor(body?: unknown, rawResponse?: core.RawResponse) {
        super({
            message: "NotFoundError",
            statusCode: 404,
            body: body,
            rawResponse: rawResponse,
        });
        Object.setPrototypeOf(this, new.target.prototype);
        if (Error.captureStackTrace) {
            Error.captureStackTrace(this, this.constructor);
        }

        this.name = this.constructor.name;
    }
}