// src/management/api/errors/PreconditionFailedError.ts
// This file was auto-generated by Fern from our API Definition.

import * as core from "../../core/index.js";
import * as errors from "../../errors/index.js";

export class PreconditionFailedError extends errors.ManagementError {
    constructor(body?: unknown, rawResponse?: core.RawResponse) {
        super({
            message: "PreconditionFailedError",
            statusCode: 412,
            body: body,
            rawResponse: rawResponse,
        });
        Object.setPrototypeOf(this, new.target.prototype);
        if (Error.captureStackTrace) {
            Error.captureStackTrace(this, this.constructor);
        }

        this.name = this.constructor.name;
    }
}