// src/management/api/errors/ServiceUnavailableError.ts
// This file was auto-generated by Fern from our API Definition.

import type * as core from "../../core";
import * as errors from "../../errors/index";

export class ServiceUnavailableError extends errors.ManagementError {
    constructor(body?: unknown, rawResponse?: core.RawResponse) {
        super({
            message: "ServiceUnavailableError",
            statusCode: 503,
            body: body,
            rawResponse: rawResponse,
        });
        Object.setPrototypeOf(this, ServiceUnavailableError.prototype);
    }
}