// src/management/api/errors/TooManyRequestsError.ts
// This file was auto-generated by Fern from our API Definition.

import type * as core from "../../core";
import * as errors from "../../errors/index";

export class TooManyRequestsError extends errors.ManagementError {
    constructor(body?: unknown, rawResponse?: core.RawResponse) {
        super({
            message: "TooManyRequestsError",
            statusCode: 429,
            body: body,
            rawResponse: rawResponse,
        });
        Object.setPrototypeOf(this, TooManyRequestsError.prototype);
    }
}